+++ routes/petshop_route.py
@@ -8,6 +8,7 @@ from models import PetshopList, PetshopModel, UpdatePetshopModel
 PETSHOP_ROUTE = Blueprint("petshop")
 collection = db['petshop']
 
+
 @PETSHOP_ROUTE.route('/get/all/petshop/', methods=['GET'])
 async def get_all_petshop(request: Request):
     all_petshops = collection.find()
@@ -60,36 +61,31 @@ async def create_petshop(request: Request):
         })
     
 
-@PETSHOP_ROUTE.route('/update/petshop/<petshop_id:str>/', methods=['PUT', 'GET'])
+@PETSHOP_ROUTE.route('/update/petshop/<petshop_id:str>/', methods=['PUT', 'PATCH', 'GET'])
 async def update_petshop(request: Request, petshop_id:str):
     try:
-        petshop_instance = UpdatePetshopModel(**request.json)
-        petshop_dict = petshop_instance.model_dump()
-        
-        collection.update_one({"_id": ObjectId(petshop_id)}, {"$set": petshop_dict})
-    
-        return response.json({"Message": "Petshop atualizado com sucesso", "Petshop atualizado": petshop_dict})
-    
-    except ValidationError as e:       
-        erros = e.errors()
-        
-        return response.json({"Description": erros[0]["type"],
-        "Message error": erros[0]['msg'],
-        })
-        
-
-@PETSHOP_ROUTE.route('patch/petshop/<petshop_id:str>/', methods=['PATCH', 'GET'])
-async def patch_petshop(request: Request, petshop_id:str):
-    try:
-        update_petshop_instance = UpdatePetshopModel(**request.json)
-        update_petshop_dict = update_petshop_instance.model_dump()
-        
-        collection.update_one({"_id": ObjectId(petshop_id)}, {"$set": update_petshop_dict})
-        
         petshop = collection.find_one({"_id": ObjectId(petshop_id)})
         
-        petshop_instance = PetshopModel(**petshop)
-        petshop_dict = petshop_instance.model_dump()
+        if petshop:
+        
+            if request.method == "PUT":
+                petshop_instance = UpdatePetshopModel(**request.json)
+                petshop_dict = petshop_instance.model_dump()
+                
+                collection.update_one({"_id": ObjectId(petshop_id)}, {"$set": petshop_dict})
+            
+            if request.method == "PATCH":
+                update_petshop_instance = UpdatePetshopModel(**request.json)
+                update_petshop_dict = update_petshop_instance.model_dump()
+                
+                collection.update_one({"_id": ObjectId(petshop_id)}, {"$set": update_petshop_dict})
+                
+                petshop = collection.find_one({"_id": ObjectId(petshop_id)})
+                
+                petshop_instance = PetshopModel(**petshop)
+                petshop_dict = petshop_instance.model_dump()
+        else:
+            return response.json({"Petshop not found"}, status=404)
         
         return response.json({"Message": "Petshop atualizado com sucesso", "Petshop atualizado": petshop_dict})
     
@@ -104,8 +100,12 @@ async def patch_petshop(request: Request, petshop_id:str):
 @PETSHOP_ROUTE.route('/delete/petshop/<petshop_id:str>/', methods=['DELETE'])
 async def delete_one_petshop(request: Request, petshop_id):
     petshop = collection.find_one({"_id": ObjectId(petshop_id)})
-    collection.delete_one(petshop)
     
-    petshop_dict = PetshopModel(**petshop).model_dump()
+    if petshop:
+        collection.delete_one(petshop)
     
-    return response.json({"Message": "Petshop deletado com sucesso", "Petshop deletado": petshop_dict})
+        petshop_dict = PetshopModel(**petshop).model_dump()
+        
+        return response.json({"Message": "Petshop deletado com sucesso", "Petshop deletado": petshop_dict})
+    else:
+        return response.json({"Petshop not found"}, status=404)
